refactor(validator): refactor gRPC server implementation
This commit refactors the gRPC server implementation in the validator
crate. The API implementation is now split into separate modules for
each endpoint.
